+++ index.go
@@ -103,18 +103,9 @@ func main() {
 	})
 
 	m.Group("/category", func(r martini.Router) {
-		//The endpoints below fetch categories/parts from a NoSQL dataset that is updated as part of our partindexing
-		r.Get("/mongo/tree", category_ctlr.GetCategoryTree)
-		r.Get("/mongo/parts/:id", category_ctlr.GetCategoryParts)
-		r.Get("/mongo/:id", category_ctlr.GetCategoryFromMongo)
-
-		r.Get("", category_ctlr.Parents)
-		r.Get("/tree", category_ctlr.Tree) // work in progress - not working yet
+		r.Get("/parts/:id", category_ctlr.GetCategoryParts)
 		r.Get("/:id", category_ctlr.GetCategory)
-		r.Post("/:id", category_ctlr.GetCategory)
-		r.Get("/:id/subs", category_ctlr.SubCategories)
-		r.Get("/:id/parts", category_ctlr.GetParts)
-		r.Post("/:id/parts", category_ctlr.GetParts)
+		r.Get("", category_ctlr.GetCategoryTree)
 	})
 
 	m.Group("/contact", func(r martini.Router) {
