+++ models/products/part.go
@@ -3,7 +3,6 @@ package products
 import (
 	"github.com/curt-labs/GoAPI/helpers/apicontext"
 	"github.com/curt-labs/GoAPI/helpers/database"
-	"github.com/curt-labs/GoAPI/helpers/redis"
 	"github.com/curt-labs/GoAPI/models/brand"
 	"github.com/curt-labs/GoAPI/models/customer"
 	"github.com/curt-labs/GoAPI/models/customer/content"
@@ -11,12 +10,6 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 	"gopkg.in/mgo.v2"
 	"gopkg.in/mgo.v2/bson"
-
-	"database/sql"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"log"
 
 	"net/url"
 	"strings"
@@ -77,15 +70,6 @@ type Class struct {
 	Image string `json:"image,omitempty" xml:"image,omitempty" bson:"image"`
 }
 
-// type Installation struct { //aka VehiclePart Table
-// 	ID          int             `json:"id,omitempty" xml:"id,omitempty" bson:"id"`
-// 	Vehicle     vehicle.Vehicle `json:"vehicle,omitempty" xml:"vehicle,omitempty" bson:"vehicle"`
-// 	Part        Part            `json:"part,omitempty" xml:"part,omitempty" bson:"part"`
-// 	Drilling    string          `json:"drilling,omitempty" xml:"v,omitempty" bson:"drilling"`
-// 	Exposed     string          `json:"exposed,omitempty" xml:"exposed,omitempty" bson:"exposed"`
-// 	InstallTime int             `json:"installTime,omitempty" xml:"installTime,omitempty" bson:"installTime"`
-// }
-
 // VehicleApplication ...
 type VehicleApplication struct {
 	Year        string `bson:"year" json:"year" xml:"year,attr"`
@@ -121,7 +105,6 @@ func (p *Part) Get(dtx *apicontext.DataContext) error {
 	err = <-databaseChan
 	close(customerChan)
 	close(databaseChan)
-	log.Print(p.Identifier)
 	return err
 }
 
@@ -239,101 +222,6 @@ func (p *Part) BindCustomer(dtx *apicontext.DataContext) CustomerPart {
 		Price:         price,
 		CartReference: ref,
 	}
-}
-
-// PartBreacrumbs
-//
-// Description: Builds out Category breadcrumb array for the current part object.
-//
-// Inherited: part Part
-// Returns: error
-func (p *Part) PartBreadcrumbs(dtx *apicontext.DataContext) error {
-	if p.ID == 0 {
-		return errors.New("Invalid Part Number")
-	}
-
-	//check redis!
-	redis_key := fmt.Sprintf("part:%d:breadcrumbs:%s", p.ID, dtx.BrandString)
-
-	data, err := redis.Get(redis_key)
-	if err == nil && len(data) > 0 {
-		if err = json.Unmarshal(data, &p.Categories); err == nil {
-			return nil
-		}
-	}
-
-	// Oh alright, let's talk with our database
-	db, err := sql.Open("mysql", database.ConnectionString())
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	partCategoryStmt, err := db.Prepare(PartCategoryStmt)
-	if err != nil {
-		return err
-	}
-	defer partCategoryStmt.Close()
-
-	lookupCategoriesStmt, err := db.Prepare(CategoriesByBrandStmt)
-	if err != nil {
-		return err
-	}
-	defer lookupCategoriesStmt.Close()
-
-	// Execute SQL Query against current ID
-	catRow := partCategoryStmt.QueryRow(p.ID)
-	if catRow == nil {
-		return errors.New("No part found for " + string(p.ID))
-	}
-
-	ch := make(chan Category)
-	go PopulateCategory(catRow, ch, dtx)
-	initCat := <-ch
-	close(ch)
-
-	// Build thee lookup
-	catLookup := make(map[int]Category)
-	rows, err := lookupCategoriesStmt.Query(dtx.BrandID)
-	if err != nil {
-		return err
-	}
-	defer rows.Close()
-
-	multiChan := make(chan []Category, 0)
-	go PopulateCategoryMulti(rows, multiChan)
-	cats := <-multiChan
-	close(multiChan)
-
-	for _, cat := range cats {
-		catLookup[cat.CategoryID] = cat
-	}
-
-	// Okay, let's put it together!
-	var categories []Category
-	categories = append(categories, initCat)
-
-	nextParentID := initCat.ParentID
-	for {
-		if nextParentID == 0 {
-			break
-		}
-		if c, found := catLookup[nextParentID]; found {
-			nextParentID = c.ParentID
-			categories = append(categories, c)
-			continue
-		}
-		nextParentID = 0
-	}
-
-	p.Categories = categories
-	if dtx.BrandString != "" {
-		go func(cats []Category) {
-			redis.Setex(redis_key, cats, redis.CacheTimeout)
-		}(p.Categories)
-	}
-
-	return nil
 }
 
 func (p *Part) GetPartByPartNumber() (err error) {
